Creates a separate group for navigation-related components
